+++ contribai/plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Plugin system for extensible analyzers and generators."""